+++ src/components/atoms/MacOSButtons.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+function MacOSButtons() {
+  return (
+    <div className="w-11 flex flex-row items-center justify-between">
+      <div className="w-[18px] h-[18px] rounded-2xl bg-white border-dark border-[3px]" />
+      <div className="w-[18px] h-[18px] rounded-2xl bg-white border-dark border-[3px]" />
+    </div>
+  );
+}
+
+export default MacOSButtons;
